Refactor ranking
