+++ src/Internal/Search/Ranking/AbstractRanker.php
@@ -10,5 +10,5 @@ abstract class AbstractRanker
      * @param array<string> $searchableAttributes
      * @param array<int, array<int, array{int, string|null}>> $termPositions
      */
-    abstract public function calculate(array $searchableAttributes, int $totalQueryTokenCount, array $termPositions): float;
+    abstract public static function calculate(array $searchableAttributes, int $totalQueryTokenCount, array $termPositions): float;
 }
